refactor text transforms for extensibility and safer rendering
- Introduce polymorphic transform data and strategy/controller boundaries.
- Extract transform control flow and group effect/rendering helpers under `ui/text_effects`.
- Preserve existing flat config/project serialization without migration.
- Add bounded global glyph geometry caches and targeted rendering optimizations.
- Test mixed transform, paste, and IME undo/redo synchronization in both layouts.
