+++ ballontranslator/ui/text_effects/__init__.py
@@ -1,0 +1,1 @@
+"""Text effect rendering isolated from the live graphics item."""
