+++ ballontranslator/utils/io_utils.py
@@ -44,6 +44,9 @@ def json_dump_nested_obj(obj, **kwargs):
     def _default(obj):
         if isinstance(obj, (np.ndarray, np.ScalarType)):
             return serialize_np(obj)
+        serializer = getattr(obj, 'to_serializable_dict', None)
+        if serializer is not None:
+            return serializer()
         return obj.__dict__
     return json.dumps(obj, default=lambda o: _default(o), ensure_ascii=False, **kwargs)
